% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliff.R
\name{cliff}
\alias{cliff}
\title{CLIFF deconvolution of drug sensitivity data}
\usage{
cliff(
  climb_output,
  drug_data,
  mutation_data = NULL,
  min.mutation = 0,
  max.em.steps = 100,
  mode = "highres",
  regularization = "none",
  cancer_pattern = "like"
)
}
\arguments{
\item{climb_output}{CLIMB output object, containing deconvoluted expression and proportions}

\item{drug_data}{data.frame containing a column named 'auc' with AUC values of drug sensitivity}

\item{mutation_data}{matrix optional parameter containing a 1/0 matrix with mutation per bulk sample}

\item{min.mutation}{integer minimum number of patient having a mutation to include it in the model}

\item{max.em.steps}{integer maximum number of steps in the EM algorithm}

\item{mode}{string 'overall' will uses the same average cell-type expression across patients. 'high-res' uses the deconvoluted expression at high-resolution produced by CLIMB}

\item{regularization}{string 'none' does not apply normalization. L2 applied ridge regularization}
}
\description{
Takes the output of CLIMB deconvolution together with drug sensitivity data as AUC values,
and optionnaly somatic mutation data that will be attributed to cancer cell-type containing
the pattern defined by 'cancer_pattern' parameter.
}
