% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliff_gene_selection.R
\name{cliff_gene_selection}
\alias{cliff_gene_selection}
\title{CLIFF gene selection}
\usage{
cliff_gene_selection(bulk_mat, drug_d, min.genes = 10)
}
\arguments{
\item{bulk_mat}{matrix containing the input bulk matrix with gene expression}

\item{drug_d}{data.frame containing a column named 'auc' with AUC values of drug sensitivity}

\item{min.genes}{integer containing the minimum number of genes to consider}
}
\description{
Perform gene selection prior to CLIFF, using multiple rounds of lasso regression.
}
