% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climb.R
\name{climb}
\alias{climb}
\title{CLIMB deconvolution}
\usage{
climb(
  sc,
  bulk,
  mode = "abundance",
  up.lim = Inf,
  lambda = 0,
  verbose = TRUE,
  cancer_pattern = "*",
  conditions = NA,
  final_res = list(),
  min_common_genes = 100,
  ratio_cell_increase = 0.02,
  n.iter.subsampling = 5,
  min.n.cells = 50,
  n.top_mean.genes = 500
)
}
\arguments{
\item{sc}{ExpressionSet object containing the scRNA-seq reference dataset}

\item{bulk}{ExpressionSet object containing the mixtures to be deconvoluted}

\item{mode}{indicate which mode to use between: \link{'all'} run bulk deconvolution of cell-type proportions - cell-type expression - and differential expression analysis (without per sample DE analysis). \link{'all+'} same as 'all' but adds per-sample DE analysis (takes long time to run). \link{'abundance'} only performs cell-type proportions inference. \link{'expression'} performs cell-type proportions together with cell-type expression prediction.}

\item{up.lim}{numeric scalar, impose a l-infinity norm to the linear model (upper bound for coefficient)}

\item{lambda}{Regularization factor}

\item{verbose}{boolean indicating whether to print message during running}

\item{cancer_pattern}{a string pattern present in all cancer cell-type. Only for these cell-types CLIMB will assume the presence of differentially expressed genes}

\item{min_common_genes}{minimum number of genes to be in common between bulk and scRNA-seq datsets}

\item{ratio_cell_increase}{percentage increase at each step of the embirical bayes procedure}

\item{n.iter.subsampling}{number of subsampling that will be performed on the single-cell reference (results from each subsample are then averaged)}

\item{min.n.cells}{minimum number of cells per cell type to subsample. If a cell type has less cells in reference, then sampling is done with replacement.}

\item{n.top_mean.genes}{number of genes used for bulk-specific gene selection.}
}
\description{
By using a scRNA-seq reference dataset (sc) and a bulk target dataset (bulk), CLIMB will generate cell-type abundance
and cell-type gene expression for each mixture given in the bulk data.
}
