% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{MSE function used by CLIFF}
\usage{
mse(tr, pr)
}
\arguments{
\item{tr}{first input vector to be compared}

\item{pr}{second input vector to be compared}
}
\description{
Callable with mse(tr, pr). Returns the mean square error between vector tr and pr.
}
