% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{RMSE function used by CLIFF}
\usage{
rmse(tr, pr)
}
\arguments{
\item{tr}{first input vector to be compared}

\item{pr}{second input vector to be compared}
}
\description{
Callable with rmse(tr, pr). Returns the root mean square error between vector tr and pr.
}
