% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid_.R
\name{sigmoid_}
\alias{sigmoid_}
\title{Sigmoid function with offset used by CLIFF}
\usage{
sigmoid_(x, a)
}
\arguments{
\item{x}{input vector to be subjected to the sigmoid function}

\item{a}{shift to apply to the sigmoid function}
}
\description{
Apply the formula 1 / 1(1 + exp(-x + a))
Callable with sigmoid_(x, a)
}
